Fixes to make it work again
Hi,
I'm wandering why melodiq doesn't working anymore, so I've checked the code.
Audiotag has changed a little the way it pass the PHPSESSID ( now through a
cookie).

I've not tried this patch because I've not a working buildhost but it will
work, I suppose ( or maybe with some little changes ).

232c232
<     runProc("curl -F uploadedfile=@" + SAMPLE2_FILE + " -F step=21
http://audiotag.info/index.php");
---
>     runProc("curl -F uploadedfile=@" + SAMPLE2_FILE + " -F step=21 -c -
http://audiotag.info/index.php");
367c367
<                 sid = extractVal("name=\"PHPSESSID\" value=\"", "\"");
---
>                 sid = extractVal("PHPSESSID       ", "\n");

Explaination:
add -c - to curl command: instruct curl to store cookie and send it to stdout.

Change extactVal strings to follow the cookie display format of curl. ( I'm
not sure if this is ok)

It's a 5 min patch, you're warned :)

Bye!