Another now correct patch to make it working again
Thanks a lot for the rebuild but, I've managed to run it on my laptop and I've
found a problem with the patch. It's not working due to PHPESSID and
extractVal.
Now, Instead of extractVal function, I'm using QRegExp to extract PHPSESSID.

This's the WORKING ( for real :) ) patch ( diff between original melodiq.cpp )